changed the folder sturcture in workato cross-adapters
