+++ packages/workato-adapter/src/filters/cross_service/netsuite/element_index.ts
@@ -13,63 +13,14 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
+import { Element, ElemID, InstanceElement, isInstanceElement, isObjectType, ObjectType } from '@salto-io/adapter-api'
 import _ from 'lodash'
-import { Element, InstanceElement, isInstanceElement, ElemID, isObjectType, ObjectType } from '@salto-io/adapter-api'
 
-export type SalesforceIndex = Record<string, Record<string, Readonly<Element>[]>>
 
 export type NetsuiteIndex = {
   scriptId: Record<string, ElemID>
   type: Record<string, Readonly<ObjectType>>
 }
-
-export type ZuoraIndex = Record<string, Readonly<Element>[]>
-
-// salesforce indexes
-
-const API_NAME = 'apiName'
-export const METADATA_TYPE = 'metadataType'
-
-const metadataType: (elem: Readonly<Element>) => string | undefined = elem => (
-  elem.annotations[METADATA_TYPE]
-)
-
-const apiName: (elem: Readonly<Element>) => string | undefined = elem => (
-  elem.annotations[API_NAME] ?? metadataType(elem)
-)
-
-export const indexSalesforceByMetadataTypeAndApiName = (
-  salesforceElements: ReadonlyArray<Readonly<Element>>,
-): SalesforceIndex => {
-  // needed in order to handle element fragments in custom objects
-  const elementsById = _.groupBy(salesforceElements, e => e.elemID.getFullName())
-
-  const mapApiNameToElem = (elements: Readonly<Element>[]): Record<string, Readonly<Element>[]> => (
-    _.mapValues(
-      _.keyBy(
-        elements.filter(e => apiName(e) !== undefined),
-        elem => apiName(elem) as string,
-      ),
-      e => elementsById[e.elemID.getFullName()],
-    )
-  )
-
-  const groupByMetadataTypeAndApiName = (
-    elements: ReadonlyArray<Readonly<Element>>
-  ): SalesforceIndex => (
-    _.mapValues(
-      _.groupBy(
-        elements.filter(e => metadataType(e) !== undefined),
-        metadataType
-      ),
-      mapApiNameToElem
-    )
-  )
-
-  return groupByMetadataTypeAndApiName(salesforceElements)
-}
-
-// netsuite indexes
 
 const CUSTOM_RECORD_TYPE = 'customrecordtype'
 const CUSTOM_RECORD_CUSTOM_FIELDS = 'customrecordcustomfields'
@@ -131,17 +82,3 @@ export const indexNetsuiteByTypeAndScriptId = (
     type: indexTypesAndFields(),
   }
 }
-
-// Zuora indexes
-
-export const indexZuoraByElemId = (
-  elements: ReadonlyArray<Readonly<Element>>
-): ZuoraIndex => _(elements)
-  .groupBy(e => e.elemID.getFullName())
-  .entries()
-  // Currently we only index objects
-  // (metadataType is our way to differ between Zuora objects to the rest of the elements)
-  .filter(([_key, group]) => group.some(e => e.annotations.metadataType !== undefined))
-  .fromPairs()
-  .mapKeys((_val, key) => key.toLowerCase())
-  .value()
