+++ packages/workato-adapter/src/filters/cross_service/reference_finders.ts
@@ -18,11 +18,17 @@ import { InstanceElement, ElemID, Value, Values } from '@salto-io/adapter-api'
 import { transformElement, TransformFunc, safeJsonStringify, setPath, extendGeneratedDependencies, FlatDetailedDependency, DependencyDirection } from '@salto-io/adapter-utils'
 import { logger } from '@salto-io/logging'
 import { strings, types, values as lowerdashValues } from '@salto-io/lowerdash'
-import { BlockBase, SupportedRecipeBlock } from './recipe_block_types'
+import { SalesforceBlock } from './salesforce/recipe_block_types'
+import { NetsuiteBlock } from './netsuite/recipe_block_types'
+import { ZuoraBlock } from './zuora_billing/recipe_block_types'
+import { BlockBase } from './recipe_block_types'
 
 const { isDefined } = lowerdashValues
 const { matchAll } = strings
 const log = logger(module)
+
+type SupportedRecipeBlock = SalesforceBlock | NetsuiteBlock | ZuoraBlock
+
 
 export type MappedReference = FlatDetailedDependency & {
   // pathToOverride is used to denote where in the element to put the reference.
